Fix compilation for datafusion core
